insomniacs/requirements.txt
Flask
configobj
mysql-connector